Prefer mounted settings over temp file

+++ utils.py
@@ -232,16 +232,16 @@ def json_load(path: Path, defaults: _JSON_T, *, merge: bool = True) -> _JSON_T:
     new_path: Path = path.with_name(f"{path.name}.new")
     combined: JsonType | None = None
 
-    if new_path.exists():
+    if path.exists():
+        with path.open('r', encoding="utf8") as file:
+            combined = _remove_missing(json.load(file, object_hook=_deserialize))
+
+    if combined is None and new_path.exists():
         try:
             with new_path.open('r', encoding="utf8") as file:
                 combined = _remove_missing(json.load(file, object_hook=_deserialize))
         except json.JSONDecodeError:
             new_path.unlink()
-
-    if combined is None and path.exists():
-        with path.open('r', encoding="utf8") as file:
-            combined = _remove_missing(json.load(file, object_hook=_deserialize))
 
     if combined is None:
         combined = dict(defaults)
